github: Auto-deploy wasm to GitHub Pages
Set up a new workflow action to build natively, then with wasm, and then
to deploy the emscripten-generated html files to a gh-pages branch. We
add a .nojekyll file to the gh-pages branch as we do not need GitHub to
bother spending time running jekyll: we can serve up the generated html
files directly.